backend/document-processing-engine/requirements.txt
celery[redis]>=5.3
redis>=5.0
psycopg2-binary>=2.9
qdrant-client>=1.7
sentence-transformers>=2.2
unstructured[pdf]>=0.12
llama-index-core>=0.10
